Add files via upload
Subarray Minimum
You are given an array (Arr) of N Distinct integers. You have to find the sum of minimum element of all the subarrays (continuous) in that array. See Sample for better understanding.

Input
The first line of input contains N, the size of the array.
The second line of input contains N space-separated integers.
Constraints
2 ≤ N ≤ 100000
1 ≤ Arr[i] ≤ 1000000000
Expected Time Complexity: O(n)
Output
The output should contain single integers, the sum of minimum element of all the subarrays in that array.