Create Catch_Me_If_You_Can.java
There are one policeman, k thieves, and one hiding spot on a coordinate line. The policeman is located at the point 0, the hiding spot is located at the point n. All the thieves are located between the policeman and the hiding spot: the i-th thief is located at the point xi (0 < xi < n). At each point, many thieves can be located.

In one second, the following happens. First, exactly one thief moves to the right by 1. If the thief reaches the hiding spot, it hides (i.e. the policeman will not be able to catch him). Then (after that the thief has finished its move) the policeman moves to the right by 1. If at the new policemans position, some thieves are located, the policeman catches them (they will not be able to move after that). The actions are performed until any thief has not been hidden or isnt caught.

In other words, the first move is made by a thief. If the thief has reached the hiding, it is saved. Then the policeman makes a move. The policeman catches the thief located at the point the policeman has reached (if the policeman has reached the hole, it catches nobody).

Each second, you can select a thief that will make a move. What is the maximum number of thieves that can reach the hiding without getting caught?

Input
The first line contains two integers n and k.
The second line contains k integers x1, x2, . . ., xk (1 ≤ xi < n) — the initial coordinates of the thieves.
Constraints
2 ≤ n ≤ 10⁹
1 ≤ k ≤ 4⋅10⁵
1 ≤ xi < n
Output
Print an integer t (t ≥ 0) — the maximum number of thieves that can reach the hiding without getting caught.

Example
Sample Input

10 6
8 7 5 4 9 4
Sample Output
3