Create Greedy_Florist.java
A group of friends want to buy a bouquet of flowers. The florist wants to maximize his number of new customers and the money he makes. To do this, he decides he will multiply the price of each flower by the number of that customers previously purchased flowers plus 1 . The first flower will be original price, (0 + 1) x original price , the next will be (1 + 1) x original price and so on.